srcs/requirements/Chat/Chat/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
coverage==7.6.9
cryptography==44.0.0
daphne==4.1.2
Django==5.0.7
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
requests==2.32.3
service-identity==24.2.0
sqlparse==0.5.1
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.6.0
zope.interface==7.2
